% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.model.interpolation.R
\name{apply.model.krige}
\alias{apply.model.krige}
\title{Apply kriging interpolation methods}
\usage{
apply.model.krige(
  x,
  soc.values,
  method = c("OK", "BK"),
  newdata,
  v.model,
  block = NULL,
  ...
)
}
\arguments{
\item{x}{spatial object with variable values}

\item{soc.values}{variable to be interpolated}

\item{method}{interpolation method: OK(ordinary kriging ) or BK (block kriging)}

\item{newdata}{new spatial data to be interpolated}

\item{v.model}{variogram model}

\item{block}{adding a block for block kriging}

\item{...}{additional arguments of krige}
}
\description{
Function to apply kriging spatial interpolation
}
\examples{
# Assuming my_grid_data is a SpatialPixelsDataFrame for prediction
# Apply kriging using the variogram model
krige_result <- apply.model.krige(my_data, "value", method = "OK", newdata = my_grid_data, v.model = vgm_model$model)

}
