% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.model.pls.R
\name{apply.model.pls}
\alias{apply.model.pls}
\title{Apply Partial Least Squares regression}
\usage{
apply.model.pls(
  spec,
  soc,
  fold.rep,
  ncomp,
  ext.rep = 5,
  int.rep = 5,
  nfold = 5,
  validation = c("CV"),
  ...
)
}
\arguments{
\item{spec}{The spectral data.}

\item{soc}{The sample ID and soil organic carbon data.}

\item{fold.rep}{A list representing the data divided into folds for cross-validation. Obtained fro the CV.folds.list function.}

\item{ncomp}{The number of components.}

\item{ext.rep}{The number of external repetitions.}

\item{int.rep}{The number of internal repetitions.}

\item{nfold}{The number of folds (should be the same as the number of folds in the `fold.rep` object).}

\item{validation}{The method of validation (currently supports cross-validation only).}

\item{...}{additional arguments passed to \link[pls]{plsr}}
}
\description{
Function for model building and validation using Partial Least Squares regression
}
\examples{
# Load data
data <- SOCmonit:::data.spec
folds <- SOCmonit:::fold.rep
spec.data <- data[, -c(1:2)]
soc.data <- data[, c(1:2)]

# Run the model
model.pls <- apply.model.pls(spec = spec.data, soc = soc.data, fold.rep = folds, ncomp = 25, ext.rep = 5, int.rep = 5, nfold = 5)
model.pls$Predictions # predictions output
model.pls$Models # individual model for each fold
model.pls$`Opt. Comp.` # optimal number of components for each model
model.pls$`Error metrics` # error metrics output
}
