% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.model.interpolation.R
\name{apply.model.vgm}
\alias{apply.model.vgm}
\title{Generate a variogram model}
\usage{
apply.model.vgm(
  soc.values,
  xcoord,
  ycoord,
  v.model,
  parameters = c(nugget = 0, p.sill, range),
  ...
)
}
\arguments{
\item{soc.values}{A numeric vector of soil organic carbon values.}

\item{xcoord}{A numeric vector of x coordinates.}

\item{ycoord}{A numeric vector of y coordinates.}

\item{v.model}{Variogram model type (as accepted by gstat::vgm).}

\item{parameters}{Initial estimates of the variogram parameters. Includes nugget, partial sill, and range.}

\item{...}{Additional arguments to be passed to gstat::vgm and gstat::variogram.}
}
\value{
A list containing the experimental variogram and the fitted variogram model.
}
\description{
This function creates an experimental variogram and fits a variogram model to the experimental data.
It's used in geostatistical studies to quantify the spatial dependence of soil organic carbon (SOC).
}
\examples{
# Assuming my_data is a dataframe with columns 'value', 'xcoord', 'ycoord'
# Apply the variogram model function
vgm_model <- apply.model.vgm(my_data$value, my_data$xcoord, my_data$ycoord, v.model = "Sph", parameters = c(0, 70000, 800))
}
