% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ASDjump.R
\name{ASDjump.plot}
\alias{ASDjump.plot}
\title{Plot sensor jump correction}
\usage{
ASDjump.plot(x1, x2, ID.point = "001a")
}
\arguments{
\item{x1}{dataframe. The dataframe created using the loadASDdata function. Needs a column named samp_loc.}

\item{x2}{dataframe. The corrected dataframe created using the loadASDdata function. Needs a column named samp_loc.}

\item{ID.point}{character: The name of the sample location to be plotted (Default: "001A")}
}
\value{
Output: 5 Plots: Extracts from ASD data at sensor jump locations before (2 plots) and after (2 plots) the correction as well as 1 plot
showing the corrected ASD data at a sample location.
}
\description{
Method to load an orthofoto and convert pixel values to reflectance based on specified camera type. The aoi is used to clip the
image to the respective plots used in the analysis.
}
\examples{
#plot results of ASD sensor jump correction
plot.ASDjump(x1=asd_data_orig, x2=asd_data_cor, ID.point ="005A")

}
