% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilePR.spec.otg.R
\name{compilePR.spec.otg}
\alias{compilePR.spec.otg}
\title{Compile predictors from on-the-go data}
\usage{
compilePR.spec.otg(x, sp.points, nb, ID.point, ID.plot, LTE, maxdist = NULL)
}
\arguments{
\item{x}{On-the-go spatial data to compile predictors from.}

\item{sp.points}{Spatial points representing the neighbors.}

\item{nb}{Number of neighbors to include for each on-the-go data point.}

\item{ID.point}{Spatial points' identifier column.}

\item{ID.plot}{ID plot of the Land Type Entity (LTE) to match against.}

\item{LTE}{LTE spatial polygon to match against.}

\item{maxdist}{Maximum distance (optional) to consider when selecting neighbors. Only neighbors within this distance will be included.}
}
\value{
A data frame containing the compiled predictors. Each row represents an on-the-go data point, and the columns contain the selected neighbors' attributes and distances.
}
\description{
This function compiles predictors from on-the-go spatial data based on spatial points and a specified neighborhood size.
}
\examples{
# Compile predictors from on-the-go data
predictors <- compilePR.spec.otg(x = onthego_data, sp.points = spatial_points, nb = 5,
                                 ID.point = "ID_point", ID.plot = "ID_plot", LTE = lte_polygon)

}
