% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilePR.UAV.R
\name{compilePR.UAV}
\alias{compilePR.UAV}
\title{Build Predictor Response Dataset}
\usage{
compilePR.UAV(x, cam_type, path1, buffering, path2, radius)
}
\arguments{
\item{x}{RasterBrick. The image loaded using loadimgdata()}

\item{cam_type}{character. The type of camera the image was recorded with. Can be "Mica" for Micasense or "Tetra" for Tetracam.}

\item{path1}{character. The file path to the shapefile containing the sampling points with a column named "SAMPLE_ID" specifying the sample location names.}

\item{buffering}{logical. If TRUE the buffered plot shapefile is given in the next parameter, if FALSE the non buffered plots are expected}

\item{path2}{character. The file path to the shapefile containing either the buffered or non buffered plots. This shapefile needs to contain a column named "PLOT_ID".}

\item{radius}{float. Defined the radius around sampling points for value extraction from the image. Define in meters or, if existing, or if existing in other map unit. Must be larger than pixel size.}
}
\value{
Output:
Predictor response dataset mean [[1]]
Predictor response dataset median [[2]]
}
\description{
Method includes functions to generate a predictor response dataset from the original or corrected image.
}
\examples{
#create predictor response datasets considering buffered plots
predrespDS <- compilePR.UAV(img_data_clip_cor, cam_type = "Mica", SPpath = "path/tofile.shp",
buffering = TRUE, plotshp_path = "path/BL_V120_crops19_20_buffer.shp", radius= 3)

pred_resp_mean_cor <- predrespDS[[1]]
pred_resp_median_cor <- predrespDS[[2]]


#create predictor response datasets without considering buffered plots
predrespDS <- compilePR.UAV(img_data_clip_cor, cam_type = "Mica", SPpath = "path/SP_(A)_V120_join.shp",
buffering = FALSE, plotshp_path = "path/BL_V120_crops19_20_buffer.shp", radius= 3)

pred_resp_mean_cor <- predrespDS[[1]]
pred_resp_median_cor <- predrespDS[[2]]

}
