% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.folds.list.R
\name{CV.folds.list}
\alias{CV.folds.list}
\title{CV fold division list}
\usage{
CV.folds.list(
  nfold,
  ext.rep,
  int.rep,
  min.dist,
  ID.point,
  soc.values,
  xcoord,
  ycoord,
  stratified = T
)
}
\arguments{
\item{nfold}{Number of folds.}

\item{ext.rep}{Number of external repetitions.}

\item{int.rep}{Number of internal repetitions.}

\item{min.dist}{Minimum distance threshold for neighbor selection in the CV.folds function.}

\item{ID.point}{Column of unique identifiers for the data points.}

\item{soc.values}{Column of the variable values.}

\item{xcoord}{Column of x-axis coordinates of the data points.}

\item{ycoord}{Column of y-axis coordinates of the data points.}

\item{stratified}{Logical indicating whether to perform stratified sampling based on quantiles.
Stratified sampling ensures that each fold is a good representative of the whole dataset.}
}
\description{
This function creates a list of training and test sets for cross-validation using the spatially aware k-fold
sampling method provided by the CV.folds function. The cross-validation process is conducted with multiple
repetitions at both external and internal levels.
}
\details{
External repetitions divide the entire dataset into k-folds. For each external repetition, internal repetitions
are performed, where each internal repetition divides the remaining data (excluding one of the external folds)
into k-folds again. This results in a nested cross-validation structure that helps provide a robust estimation
of model performance and stability across different partitionings of the data.

Each iteration of the internal repetitions produces a new fold assignment (named as "fold.int [iteration]")
added to the data. The resulting list contains entries for each external repetition. Each entry in the list is a
data frame with the original data, fold assignments from the external repetition (named "fold"), and fold
assignments from all internal repetitions.
}
\examples{
# Generate CV fold division list
cv_list <- CV.folds.list(nfold = 5, ext.rep = 3, int.rep = 2, min.dist = 0.5,
                         ID.point = df$ID, soc.values = df$SOC,
                         xcoord = df$x, ycoord = df$y, stratified = T)

}
