% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.folds.R
\name{CV.folds}
\alias{CV.folds}
\title{Perform K-fold sampling with spatial neighbor constraints}
\usage{
CV.folds(x, min.dist, nfolds, stratified = TRUE, xcoord, ycoord)
}
\arguments{
\item{x}{Numeric vector or matrix of data.}

\item{min.dist}{Numeric value specifying the minimum distance threshold for neighbor selection. Pairs of samples
closer than this threshold are considered neighbors and placed within the same fold.}

\item{nfolds}{Integer specifying the number of folds for sampling.}

\item{stratified}{Logical indicating whether to perform stratified sampling based on quantiles of \code{x}.
Stratified sampling ensures that each fold is a good representative of the whole dataset.}

\item{xcoord}{Numeric vector of x-coordinates corresponding to the spatial locations of the data.}

\item{ycoord}{Numeric vector of y-coordinates corresponding to the spatial locations of the data.}
}
\value{
A data frame with the sampled data and fold assignments.
}
\description{
This function performs k-fold sampling while considering spatial autocorrelation.
}
\details{
The function first computes a spatial distance matrix using the provided coordinates. For each pair of
  samples, if the distance is below the provided minimum distance threshold, the samples are marked as neighbors.
  The function then creates k-folds either randomly or in a stratified manner (if specified). It then iteratively
  reassigns the folds of the neighbor pairs in such a way that they belong to the same fold, while ensuring that
  each fold has at least one sample. If the function cannot achieve this configuration due to the number of close
  neighbors being larger than the fold size, it stops and raises an error. If there are no neighbors below the
  minimum distance, the function proceeds with the created folds without reassigning any samples.
}
\examples{
x <- 1:10
xcoord <- runif(10)
ycoord <- runif(10)
CV.folds(x, min.dist = 0.5, nfolds = 5, stratified = TRUE, xcoord, ycoord)
}
