% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.results.plot.R
\name{pls.results.plot}
\alias{pls.results.plot}
\title{Plot PLS Regression Results}
\usage{
pls.results.plot(model.pls)
}
\arguments{
\item{model.pls}{The PLS regression model object.}
}
\value{
None
}
\description{
Function to visualize the results of PLS regression.
}
\examples{
# Assuming model_pls is a PLS regression model object
plot.pls.results(model_pls)

}
