% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.spec.R
\name{pp.spec}
\alias{pp.spec}
\title{Preprocess Spectral data}
\usage{
pp.spec(
  spectra,
  columns = 1:ncol(spectra),
  pp.method = "snv",
  wav = NULL,
  type = NULL,
  interpol = NULL,
  new.wav = NULL,
  sigma = NULL,
  targetnorm = NULL,
  m = 1,
  p = 1,
  w = 5,
  fwhm = NULL,
  bins = NULL,
  bin.size = NULL,
  ...
)
}
\arguments{
\item{spectra}{The input spectral data. Can be a matrix or data frame.}

\item{columns}{The columns of the spectral data to be processed.}

\item{pp.method}{The preprocessing method to apply. Available options: "cr" (continuum removal),
"sg" (Savitzky-Golay smoothing), "dt" (detrend), "gd" (gap derivative), "ma" (moving average),
"msc" (multiplicative scatter correction),"snv" (standard normal variate), "b" (binning),
"rs" (resample), "rs2" (resample2), "bs" (block scaling), "bn" (block normalization).}

\item{wav}{Wavelength vector used for preprocessing methods that require it (cr, dt, rs, rs2).}

\item{type}{Type of continuum removal method (cr). Available options: "rubberband" or "modpoly".}

\item{interpol}{Interpolation method for resampling (rs). Available options: "linear" or "spline".}

\item{new.wav}{Wavelength vector for resampling (rs, rs2).}

\item{sigma}{Sigma value for block scaling (bs).}

\item{targetnorm}{Target normalization value for block normalization (bn).}

\item{m}{Polynomial order for Savitzky-Golay smoothing (sg).}

\item{p}{Number of points to use for Savitzky-Golay smoothing (sg).}

\item{w}{Width of the Savitzky-Golay window (sg, ma).}

\item{fwhm}{Full Width at Half Maximum for resample2 method (rs2).}

\item{bins}{Number of bins for binning method (b).}

\item{bin.size}{Size of each bin for binning method (b).}

\item{...}{Additional parameters to be passed to the respective prospectr functions.}
}
\value{
Processed spectral data with the specified preprocessing applied.
}
\description{
Applies preprocessing methods to spectral data using prospectr package see \link[prospectr]{prospectr}
}
\examples{
# Example 1: Standard Normal Variate (snv) preprocessing
pp.spec(spectra, pp.method = "snv")

# Example 2: Savitzky-Golay smoothing
pp.spec(spectra, pp.method = "sg", m = 2, p = 1, w = 5)

# Example 3: Resample with interpolation
pp.spec(spectra, pp.method = "rs", interpol = "linear", new.wav = new_wav, wav = wav)

}
