% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.ASD.R
\name{rd.ASD}
\alias{rd.ASD}
\title{Read ASD data and create dataframe}
\usage{
rd.ASD(path, year, jumpcorr = TRUE, corrmethod = "diff")
}
\arguments{
\item{path}{character. The path to the directory where the ASD data (without header) is saved}

\item{year}{character. The year that the ASD data was recorded. Currently only 2018 and 2019 are supported.}

\item{jumpcorr}{logial. If TRUE (default), sensor jump correction is conducted.}

\item{corrmethod}{character. Can be difference ("diff" - default) or multiplicative ("multi") approach.}
}
\value{
Output: If jumpcorr=F: Dataframe with original ASD data. If jumpcorr=T: List of two dataframes. ASD data before [[1]] and after [[2]] sensor jump correction.
}
\description{
Reads the ASD from individual .txt files into a dataframe and returns the original and corrected ASD dataframes as a list.
}
\examples{
#read ASD data and conduct sensor jump correction
as_data <- rd.ASD(path="Path/todirectory", year="2018", jumpcorr=TRUE, corrmethod="diff")
#asd data without correction
asd_data_orig <- asd_data[[1]]
#asd data with correction
asd_data_cor <- asd_data[[2]]

#read ASD dara without sensor jump correction
as_data <- rd.ASD(path="Path/todirectory", year="2018", jumpcorr=FALSE)

}
