% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.LTE.R
\name{rd.LTE}
\alias{rd.LTE}
\title{Read LTE shapefile}
\usage{
rd.LTE(file, ID.plot = NULL, ...)
}
\arguments{
\item{file}{Location of the shapefile.}

\item{ID.plot}{Column name or index with the plot ID.}
}
\description{
Read a Long-Term Experiment (LTE) shapefile and return the data.
}
\examples{
out.LTE<-  rd.LTE(system.file("extdata/LTE.shp", package="SOCmonit"), ID.plot="ID_plot")
}
