% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.soildata.R
\name{rd.soildata}
\alias{rd.soildata}
\title{Read soil data}
\usage{
rd.soildata(
  file,
  type = c("csv", "text"),
  soc.values,
  ID.sample = NULL,
  ID.point,
  average_by_point = FALSE,
  ...
)
}
\arguments{
\item{file}{File location.}

\item{type}{File type indicator.
- "csv" indicates that the file is in CSV  format.
- "text" indicates that the file is in plain text format.}

\item{soc.values}{Column name of soc values.}

\item{ID.sample}{Column name of Sample ID.}

\item{ID.point}{Column name of point ID (location number).}

\item{average_by_point}{Logical indicating whether to average the data by ID.point.}

\item{...}{Arguments to be passed to read.csv and read.table functions.}
}
\value{
A data frame containing the soil data.
}
\description{
Read soil data from external files.
}
\examples{
# Read soil data from a CSV file without averaging
out.socdata <- rd.soildata(system.file("extdata/socdata.csv",package="SOCmonit"), type = "csv", soc.values = "C",
                          ID.sample = "Code", ID.point = "ID", average_by_point = FALSE)

# Read soil data from a CSV file with averaging
out.socdata_avg <- rd.soildata(system.file("extdata/socdata.csv",package="SOCmonit"), type = "csv", soc.values = "C",
                              ID.sample = NULL, ID.point = "ID", average_by_point = TRUE)

}
