% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.spec.otg.R
\name{rd.spec.otg}
\alias{rd.spec.otg}
\title{Read on-the-go spectra data}
\usage{
rd.spec.otg(
  file,
  type = c("csv", "text"),
  xcoord,
  ycoord,
  spec.columns = NULL,
  projection = NULL,
  non.duplicates = FALSE,
  to.UTM = FALSE,
  UTM.proy = NULL,
  ...
)
}
\arguments{
\item{file}{A character string of the file location.}

\item{type}{A character string specifying the file type, defaults to csv.}

\item{xcoord}{column name containing x axis coordinates}

\item{ycoord}{column name containing y axis coordinates}

\item{spec.columns}{define column range of spectral data}

\item{projection}{define projection (CRS arguments)}

\item{non.duplicates}{logical. if TRUE will remove duplicates}

\item{to.UTM}{logical. if TRUE will transform to UTM}

\item{UTM.proy}{add CRS argument for UTM coordinates}

\item{...}{arguments to be passed to \link[utils]{read.csv} and \link[utils]{read.table}}
}
\value{
An object of SpatialPointsDataFrame and a plot with point locations
}
\description{
Read on-the-go spectral data from a file
}
\details{
Allows to define coordinates system and transform degrees to UTM
}
\examples{
#read a csv file
out.linespec<- rd.spec.otg(system.file("extdata/spec.onthego.csv", package="SOCmonit"),
type="csv",xcoord="Long",ycoord="Lat",spec.columns = 20:400, non.duplicates = F)

#remove duplicates
out.linespec<- rd.spec.otg(system.file("extdata/spec.onthego.csv", package="SOCmonit"),
                         type="csv",xcoord="Long",ycoord="Lat",
                          spec.columns = 20:400, non.duplicates =T)
#convert Lat/Long to UTM
out.linespec<- rd.spec.otg(system.file("extdata/spec.onthego.csv", package="SOCmonit"),
                         type="csv",xcoord="Long",ycoord="Lat", spec.columns = 20:400, non.duplicates =T,
                         projection=("+proj=longlat +datum=WGS84"), to.UTM = T, UTM.proy="+init=epsg:25833")

}
