% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.spoints.R
\name{rd.spoints}
\alias{rd.spoints}
\title{Read and visualize Sampling Points}
\usage{
rd.spoints(
  file,
  type = c("csv", "text"),
  xcoord,
  ycoord,
  ID.point = NULL,
  projection = NULL,
  ...
)
}
\arguments{
\item{file}{File location of the data.}

\item{type}{File type: "csv" for CSV files, "text" for text files.}

\item{xcoord}{Column name of the x-axis coordinates.}

\item{ycoord}{Column name of the y-axis coordinates.}

\item{ID.point}{Column name of the point ID (optional).}

\item{projection}{Projection information (CRS arguments) for spatial data (optional).}

\item{...}{Additional arguments to be passed to read.csv and read.table functions.}
}
\value{
A SpatialPointsDataFrame object representing the sampling points.
}
\description{
Read sampling points coordinates from a file and visualize them on a scatter plot.
}
\examples{
# Read sampling points from a CSV file and plot them
rd.spoints("points.csv", type = "csv", xcoord = "longitude", ycoord = "latitude", ID.point = "ID", projection = "+proj=longlat +datum=WGS84")

}
