% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.srf.R
\name{rd.srf}
\alias{rd.srf}
\title{Load SRF data from Excel file and create dataframe}
\usage{
rd.srf(path, cam_type)
}
\arguments{
\item{path}{character. The path to the directory where the SRF file(s) is saved. Format Tetra: 490FS10-25 92778 X242-241 59116 23^C 0^A RANDOM^POL_1.csv, Format Mica: RedEdge_3_Filters_srs-3.ods.}

\item{cam_type}{character. The type of camera the image was recorded with. Can be "Mica" for Micasense or "Tetra" for Tetracam}
}
\description{
Loads the SRF data from an Excel file and creates a dataframe. The final shape of the dataframe is the same for
both camera systems but requires different processing steps.
}
\examples{
#read spectral response functions for Micasense
srf_mica <- rd.srf(path="Path/todirectory", cam_type="Mica")

#' #read spectral response functions for Tetracam
srf_tetra <- rd.srf(path="Path/todirectory", cam_type="Tetra")

}
