% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.UAV.R
\name{rd.UAV}
\alias{rd.UAV}
\title{Read image data and option for clipping}
\usage{
rd.UAV(path, cam_type, aoipath)
}
\arguments{
\item{path}{character. The file path to the image that is to be loaded}

\item{cam_type}{character. The type of camera the image was recorded with. Can be "Mica" for Micasense or "Tetra" for Tetracam}

\item{aoipath}{character. The file path to the aoi that is used to clip the image}
}
\value{
Raster dataset as a whole or clipped.
}
\description{
Method to read an orthofoto and convert pixel values to reflectance based on specified camera type. The aoi is used to clip the
image to the respective plots used in the analysis.
}
\examples{
#read UAV image and clip to area of interest
img_data_clip <- rd.UAV(path="Path/tofile.tif", cam_type="Mica", aoipath="Path/tofile.shp")



}
