% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.point.spec.R
\name{report.point.spec}
\alias{report.point.spec}
\title{Generate a report summarizing the results of preprocessing and modeling point spectral data for SOC prediction.}
\usage{
report.point.spec(
  output_dir,
  output_file,
  spectra_data = NULL,
  test_results = NULL,
  soil_data = NULL,
  sampling_points = NULL,
  pp_data = NULL,
  fold_list = NULL,
  model_pls = NULL
)
}
\arguments{
\item{output_dir}{The directory where the report will be saved.}

\item{output_file}{The name of the output file without the file extension.}

\item{spectra_data}{The spectral data with corrected values.}

\item{test_results}{Results of testing the spectral data.}

\item{soil_data}{The soil data for SOC prediction.}

\item{sampling_points}{The scatter plot of sampling points.}

\item{pp_data}{The preprocessed ASD data.}

\item{fold_list}{The k-fold partitions for cross-validation.}

\item{model_pls}{The results of the PLS modeling.}
}
\value{
The path to the generated PDF report.
}
\description{
This function generates a report summarizing the results obtained from using the SOCmonit package for preprocessing and modeling point spectral data for SOC prediction. The report covers various steps, including reading and testing point spectral data, reading soil data, preprocessing ASD data, creating a k-fold list for cross-validation, applying PLS modeling, and visualizing the results.
}
\details{
If TinyTeX is not installed on the system, this function will automatically install it to ensure proper LaTeX compilation for generating the report.
}
\examples{
# Generate a report with all sections
report.point.spec(output_dir = "report", output_file = "summary",
                  spectra_data = spectra_data, test_results = test_results,
                  soil_data = soil_data, sampling_points = sampling_points,
                  pp_data = pp_data, fold_list = fold_list, model_pls = model_pls)

# Generate a report with only specific sections
report.point.spec(output_dir = "report", output_file = "summary",
                  spectra_data = spectra_data, soil_data = soil_data)

}
