% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.UAV.R
\name{report.uav}
\alias{report.uav}
\title{Generate a report summarizing the results of UAV spectral data analysis.}
\usage{
report.uav(
  output_dir,
  output_file,
  UAV_data = NULL,
  asd_data = NULL,
  srf_data = NULL,
  processing_output = NULL,
  predrespDS = NULL,
  Cam_type = c("Mica", "Tetra")
)
}
\arguments{
\item{output_dir}{The directory where the report will be saved.}

\item{output_file}{The name of the output file without the file extension.}

\item{UAV_data}{The UAV image data.}

\item{asd_data}{The ASD data.}

\item{srf_data}{The SRF data.}

\item{processing_output}{The output from the image processing step.}

\item{predrespDS}{The predictor response dataset.}

\item{Cam_type}{The type of camera used (Mica or Tetra).}
}
\value{
The path to the generated PDF report.
}
\description{
This function generates a report summarizing the results obtained using the SOCmonit package for application of key functions for pre-processing and comparing field spectrometer data with drone data.
}
\details{
If TinyTeX is not installed on the system, this function will automatically install it to ensure proper LaTeX compilation for generating the report.
}
\examples{
# Generate a report with all sections
report.uav(output_dir = "report", output_file = "summary",
           UAV_data = UAV_data, asd_data = asd_data, srf_data = srf_data,
           processing_output = processing_output, predrespDS = predrespDS, Cam_type = "Mica")

# Generate a report with only specific sections
report.uav(output_dir = "report", output_file = "summary",
           UAV_data = UAV_data, asd_data = asd_data)

}
