% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.spec.otg.R
\name{test.spec.otg}
\alias{test.spec.otg}
\title{Test on-the-go spectral data}
\usage{
test.spec.otg(
  spectra,
  columns = 1:ncol(spectra),
  ID.plot = NULL,
  ID.otg = NULL,
  outliers_by_id_plot = TRUE,
  ...
)
}
\arguments{
\item{spectra}{An object containing spectral data. It can be either a matrix or an S4 object (e.g., SpatialPointsDataFrame).}

\item{columns}{A vector specifying the indices or names of the columns in the \code{spectra} object that contain the spectral data.}

\item{ID.plot}{A vector specifying the indices or names of the columns in the \code{spectra} object that contain the plot IDs (optional).}

\item{ID.otg}{A vector specifying the indices or names of the columns in the \code{spectra} object that contain the outlier group IDs (optional).}

\item{outliers_by_id_plot}{A logical value indicating whether to perform outlier detection separately for each ID.plot (default is TRUE).}

\item{...}{Additional arguments to be passed to the \code{mvoutlier::pcout} function.}
}
\value{
A list containing the results of outlier detection. The structure of the output depends on the \code{outliers_by_id_plot} argument.
If \code{outliers_by_id_plot} is FALSE, the output will contain general results, including a list of no outliers and outliers.
The no outliers will be returned as a SpatialPointsDataFrame object with the same coordinate reference system (CRS) as the input \code{spectra} object.
The outliers will be returned as a dataframe.
If \code{outliers_by_id_plot} is TRUE, the output will contain individual results, including a list of no outliers and outliers for each ID.plot.
The no outliers (for general or each ID.plot) will be returned as a SpatialPointsDataFrame object(s) with the same CRS as the input \code{spectra} object.
The outliers (for general or each ID.plot) will be returned as a dataframe(s).
}
\description{
This function performs outlier detection on spectral data, either in a general or separated by Plot ID.
}
\examples{
# Example 1: General outlier detection
data <- matrix(rnorm(100), ncol = 10)  # Sample spectral data
result <- test.spec.otg(data)  # Perform general outlier detection
print(result)

# Example 2: Individual outlier detection by ID.plot
data <- data.frame(ID.plot = c(1, 1, 2, 2, 2, 3, 3), Spectra = rnorm(7))
result <- test.spec.otg(data, ID.plot = "ID.plot", outliers_by_id_plot = TRUE)
print(result)

}
