% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.spec.pts.R
\name{test.spec.pts}
\alias{test.spec.pts}
\title{Test point spectral data}
\usage{
test.spec.pts(
  spectra,
  columns = 1:ncol(spectra),
  ID.point,
  ID.sample = NULL,
  outliers_by_id_point = TRUE,
  include_average = FALSE,
  ...
)
}
\arguments{
\item{spectra}{The input data containing the spectra.}

\item{columns}{A vector specifying the columns of the spectra to be considered.}

\item{ID.point}{The column index or name indicating the identifier for each spectrum.}

\item{ID.sample}{The column index or name indicating the identifier for each sample. Defaults to NULL.}

\item{outliers_by_id_point}{Logical value indicating whether to calculate outliers separately for each ID.point.
If TRUE, outliers will be calculated for each ID.point individually.
If FALSE, general outliers for all spectra will be calculated.}

\item{include_average}{Logical value indicating whether to include the average of non-outlier spectra by ID.point.}

\item{...}{Additional arguments to be passed to the \code{mvoutlier::pcout} function.}
}
\value{
A list containing the calculated results.
}
\description{
Identify and remove outliers from spectral data based on the PCOut function
}
\details{
This function calculates outliers and summary statistics for a given set of spectra.

The function calculates outliers using the \code{mvoutlier::pcout} function based on the provided spectra data.
If \code{outliers_by_id_point} is set to TRUE, it separates the spectra into individual ID.points and calculates outliers
for each ID.point separately. The outliers are stored in the \code{Individual_results} list, which contains two elements:
\code{No_outliers} and \code{Outliers}. The \code{No_outliers} element contains the non-outlier spectra for each ID.point,
and the \code{Outliers} element contains the outlier spectra for each ID.point.

If \code{outliers_by_id_point} is set to FALSE, it calculates general outliers for all spectra. The general outliers are
stored in the \code{General_results} list, which contains two elements: \code{No_outliers} and \code{Outliers}. The
\code{No_outliers} element contains the non-outlier spectra, and the \code{Outliers} element contains the outlier spectra.

Additionally, if \code{include_average} is set to TRUE, it calculates the average of the non-outlier spectra by ID.point
and stores it in the \code{Average_no_outliers} element of the output list.
}
\examples{
# Example with a data frame
data <- data.frame(ID.point = c(1, 1, 2, 2, 3, 3),
                   ID.sample = c(1, 2, 1, 2, 1, 2),
                   spec = c(1.2, 1.3, 1.1, 1.2, 0.9, 1.0))
result <- test.spec.pts(data, columns = "spec", ID.point = "ID.point", outliers_by_id_point = TRUE)

}
