% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{thresholds}
\alias{thresholds}
\title{NDVI and BVIS thresholds}
\usage{
thresholds(x)
}
\arguments{
\item{x}{RasterBrick. The image loaded using loadimgdata()}
}
\value{
Function returns a scatter plot of
NDVI and BVIS as well as individual historgrams for both.
}
\description{
Method to generate NDVI and BVIS indices and determine threshold values used for creating masks.
}
\examples{
#plot values for NDVI and BVIS calculated from input image
thresholds(x=UAV_img_clipped)

}
