% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.UAV_UAV.ASDcor.R
\name{UAV_UAV.ASDcor.plot}
\alias{UAV_UAV.ASDcor.plot}
\title{Comparison plot}
\usage{
UAV_UAV.ASDcor.plot(x1, x2, x3, x4, cam_type, ID.point = "005A")
}
\arguments{
\item{x1}{dataframe. The ASD data created using the loadASDdata function or in the specified format.}

\item{x2}{dataframe. The dataframe holding sample point data from the UAV data (output of the pp.UAVprocessing function)}

\item{x3}{dataframe. asd data after srf translation}

\item{x4}{dataframe. values extracted from the corrected image}

\item{cam_type}{character. The type of camera the image was recorded with. Can be "Mica" for Micasense or "Tetra" for Tetracam}

\item{ID.point}{character. The name of the sample location to be plotted (Default: "005A")}
}
\description{
Method to compare (corrected) ASD data, camera data and corrected camera data at all sample locations
}
\examples{
#plot comparison of corrected ASD data camera data and corrected camera data for each sample location
plot.UAV_UAV.ASDcor(x1=asd_data_cor, x2=cam_sample_pts, x3=asd_cam, x4=corimg_samp_pts, cam_type="Mica", ID.point="001A")


}
