% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.UAVspec.pts.R
\name{UAVspec.pts.plot}
\alias{UAVspec.pts.plot}
\title{Comparison plot}
\usage{
UAVspec.pts.plot(x1, x2, x3, ID.point = "040A", cam_type, xmin, xmax)
}
\arguments{
\item{x1}{dataframe. The ASD data created using the loadASDdata function or in the specified format.}

\item{x2}{dataframe. The dataframe holding spectral UAV data at the sampling points (output of the pp.UAVprocessing function)}

\item{x3}{asd data after srf translation}

\item{ID.point}{character. The name of the sample location to be plotted (Default: "040A")}

\item{cam_type}{character. The type of camera the image was recorded with. Can be "Mica" for Micasense or "Tetra" for Tetracam}

\item{xmin}{numeric. Minimum value displayed on the x-axis}

\item{xmax}{numeric. Maximum value displayed on the x-axis}
}
\value{
Output: Graph, boxplot for difference (asd_cam vs cam_sample_pts) and ratio (asd_cam vs cam_sample_pts)
}
\description{
Method to visualize the difference between the image values and the srf corrected ASD data that are used to correct the image values.
}
\examples{
#plot difference between the image values and the srf corrected ASD data
plot.UAVspec.pts(x1=asd_data_cor, x2=cam_sample_pts, x3=asd_cam, ID.point = "003A", cam_type = "mica", xmin = 350, xmax = 2000)

}
