% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.UAVsrf.R
\name{UAVsrf.plot}
\alias{UAVsrf.plot}
\title{SRF Plot}
\usage{
UAVsrf.plot(x, cam_type, xmin, xmax)
}
\arguments{
\item{x}{SRF data generated using the loadSRFdata function}

\item{cam_type}{character. The type of camera the image was recorded with. Can be "Mica" for Micasense or "Tetra" for Tetracam}

\item{xmin}{numeric. Minimum value displayed on the x-axis}

\item{xmax}{numeric. Maximum value displayed on the x-axis}
}
\description{
Method to visualize spectral response functions (SRF) used in the further analysis.
}
\examples{
#plot the loaded spectral response function data for MicaSense 
plot.UAVsrf(x=srf_mica, cam_type="Mica", xmin = 350, xmax = 1000)

#plot the loaded spectral response function data for Tetracam 
plot.UAVsrf(x=srf_tetra, cam_type="Tetra", xmin = 350, xmax = 1000)

}
